refactor: Update weather icon to feature
