+++ src/features/discovery/ui/weather-icon.tsx
