+++ src/widgets/discovery/ui/weekly-weather.tsx
@@ -1,6 +1,6 @@
 import type { WeeklyWeather } from '@/entities/discovery/model/model';
+import WeatherIcon from '@/features/discovery/ui/weather-icon';
 import { cn } from '@/shared/lib';
-import WeatherIcon from './weather-icon';
 
 interface WeeklyWeatherProps extends WeeklyWeather {
   isToday: boolean;
